requirements.txt
streamlit
torch 
torchvision 
pillow 
huggingface-hub
timm
plotly
torchsummary